backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.5,<6.0
django-cors-headers>=4.9,<5.0
psycopg[binary]>=3.2,<4.0
weasyprint>=66.0,<67.0
